+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## v73.0.0-SNAPSHOT - unreleased
+## v72.5.0 - 2025-04-14
 
 ### 🎁 New Features